requirements.txt
# install the following packages
csv 
biopython
tqdm
pandas
concurrent.futures
cupy